Bit of PoC cleanup
