+++ pages/wizard.vue
@@ -7,123 +7,139 @@
   >
     <section class="pf-c-page__main-section pf-m-light">
       <div class="pf-c-content">
-        <h1>New Project</h1>
+        <h1>Residency Wizard</h1>
         <p>Set up a new project to request resources</p>
       </div>
     </section>
-    <section class="pf-c-page__main-wizard">
-      <div class="pf-c-wizard pf-m-in-page fill-wizard">
-        <button
-          aria-label="Wizard Header Toggle"
-          class="pf-c-wizard__toggle"
-          aria-expanded="false"
-        >
-          <ol class="pf-c-wizard__toggle-list">
-            <li class="pf-c-wizard__toggle-list-item">
-              <span class="pf-c-wizard__toggle-num">2</span>Point of Contact
-              <!-- <i
-                class="fas fa-angle-right pf-c-wizard__toggle-separator"
-                aria-hidden="true"
-              >
-              </i> -->
-            </li>
-            <!-- <li class="pf-c-wizard__toggle-list-item">Substep B</li> -->
-          </ol>
-          <i
-            class="fas fa-caret-down pf-c-wizard__toggle-icon"
-            aria-hidden="true"
-          >
-          </i>
-        </button>
-        <div class="pf-c-wizard__outer-wrap">
-          <div class="pf-c-wizard__inner-wrap">
-            <nav class="pf-c-wizard__nav" aria-label="Steps">
-              <ol class="pf-c-wizard__nav-list">
-                <li class="pf-c-wizard__nav-item">
-                  <a href="#" class="pf-c-wizard__nav-link">
-                    Basic Information
-                  </a>
-                </li>
-                <li class="pf-c-wizard__nav-item">
-                  <a href="#" class="pf-c-wizard__nav-link pf-m-current">
-                    Point of Contact
-                  </a>
-                  <!-- <ol class="pf-c-wizard__nav-list">
-                    <li class="pf-c-wizard__nav-item">
-                      <a href="#" class="pf-c-wizard__nav-link">Substep A</a>
-                    </li>
-                    <li class="pf-c-wizard__nav-item">
-                      <a
-                        href="#"
-                        class="pf-c-wizard__nav-link pf-m-current"
-                        aria-current="page"
-                      >
-                        Substep B
-                      </a>
-                    </li>
-                    <li class="pf-c-wizard__nav-item">
-                      <a href="#" class="pf-c-wizard__nav-link">Substep C</a>
-                    </li>
-                  </ol> -->
-                </li>
-                <li class="pf-c-wizard__nav-item">
-                  <a href="#" class="pf-c-wizard__nav-link">
-                    OpenShift Cluster
-                  </a>
-                </li>
-                <li class="pf-c-wizard__nav-item">
-                  <a href="#" class="pf-c-wizard__nav-link">
-                    Tools &amp; Services
-                  </a>
-                </li>
-                <li class="pf-c-wizard__nav-item">
-                  <a href="#" class="pf-c-wizard__nav-link">
-                    Users
-                  </a>
-                </li>
-                <li class="pf-c-wizard__nav-item">
-                  <a
-                    href="#"
-                    class="pf-c-wizard__nav-link pf-m-disabled"
-                    aria-disabled="true"
-                    tabindex="-1"
-                  >
-                    Review
-                  </a>
-                </li>
-              </ol>
-            </nav>
-            <div class="pf-c-wizard__main">
-              <div class="pf-c-wizard__main-body">
-                <p>Wizard content goes here</p>
-              </div>
-            </div>
+    <Wizard :navigation="navItems" initialState="basic-information">
+      <template slot="basic-information">
+        <form novalidate class="pf-c-form">
+          <div class="pf-c-form__group">
+            <label class="pf-c-form__label" for="customer-name">
+              <span class="pf-c-form__label-text">Customer</span>
+              <span class="pf-c-form__label-required" aria-hidden="true">
+                &#42;
+              </span>
+            </label>
+            <input
+              class="pf-c-form-control"
+              required
+              type="text"
+              id="customer-name"
+              name="customer-name"
+              v-model="customer"
+            />
+            <p class="pf-c-form__helper-text" aria-live="polite">
+              The customer name to use for this residency
+            </p>
           </div>
-          <footer class="pf-c-wizard__footer">
-            <button class="pf-c-button pf-m-primary" type="submit">Next</button>
-            <button class="pf-c-button pf-m-secondary" type="button">
-              Back
-            </button>
-            <button class="pf-c-button pf-m-link" type="button">Cancel</button>
-          </footer>
-        </div>
-      </div>
-    </section>
+          <div class="pf-c-form__group">
+            <label class="pf-c-form__label" for="customer-location">
+              <span class="pf-c-form__label-text">Location</span>
+              <span class="pf-c-form__label-required" aria-hidden="true">
+                &#42;
+              </span>
+            </label>
+            <input
+              class="pf-c-form-control"
+              required
+              type="email"
+              id="customer-location"
+              name="customer-location"
+              v-model="location"
+            />
+            <p class="pf-c-form__helper-text" aria-live="polite">
+              Where the residency will be held
+            </p>
+          </div>
+          <div class="pf-c-form__group">
+            <label class="pf-c-form__label" for="simple-form-number">
+              <span class="pf-c-form__label-text">Dates</span>
+            </label>
+            <br />
+            <template>
+              <client-only>
+                <date-picker
+                  v-model="dates"
+                  input-class="pf-c-form-control"
+                  range
+                />
+              </client-only>
+            </template>
+            <p class="pf-c-form__helper-text" aria-live="polite">
+              When the residency will be held
+            </p>
+          </div>
+        </form>
+      </template>
+      <template slot="point-of-contact">
+        point of contact stuff
+      </template>
+      <template slot="openshift-cluster">
+        openshift stuff
+      </template>
+      <template slot="tools-and-services">
+        tools and stuff
+      </template>
+      <template slot="users">
+        users and stuff
+      </template>
+    </Wizard>
   </main>
 </template>
 
 <script>
+import Wizard from '~/components/patternfly/Wizard.vue'
+
+const navItems = [
+  {
+    key: 'basic-information',
+    displayName: 'Basic Information'
+  },
+  {
+    key: 'point-of-contact',
+    displayName: 'Point of Contact'
+  },
+  {
+    key: 'openshift-cluster',
+    displayName: 'OpenShift Cluster'
+  },
+  {
+    key: 'tools-and-services',
+    displayName: 'Tools & Services'
+  },
+  {
+    key: 'users',
+    displayName: 'Users'
+  }
+]
+
 export default {
-  components: {}
+  components: {
+    Wizard
+  },
+  data: () => {
+    return {
+      navItems,
+      dates: '',
+      customer: '',
+      location: ''
+    }
+  }
 }
 </script>
 
-<style scoped>
-.fill-wizard {
-  min-height: calc(100vh - 76px - 112px);
-  /*                ^ 100% window height
-                            ^ header height
-                                    ^ title height
-  */
+<style>
+.pf-c-form-control {
+  width: 100%;
+}
+.mx-datepicker {
+  width: 100%;
+}
+.mx-input-wrapper {
+  width: 100%;
+}
+td.cell {
+  text-align: center;
 }
 </style>
